refactor : Pin -> Pins 객체명 변경
