--- src/test/java/bowling/domain/PinTest.java
@@ -1,35 +1,0 @@
-package bowling.domain;
-
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.assertThatExceptionOfType;
-
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.ValueSource;
-
-public class PinTest {
-
-	@ParameterizedTest
-	@ValueSource(ints = {1, 2, 3, 4, 5, 6, 7, 8, 9, 10})
-	@DisplayName("핀 생성 성공")
-	void create(int pin) {
-		assertThat(new Pin(pin)).isInstanceOf(Pin.class);
-	}
-
-	@ParameterizedTest
-	@ValueSource(ints = {-1, -2})
-	@DisplayName("핀 수가 최소 수 보다 작을 경우")
-	void minimumException(int pin) {
-		assertThatExceptionOfType(IllegalArgumentException.class)
-			.isThrownBy(() -> new Pin(pin));
-	}
-
-	@ParameterizedTest
-	@ValueSource(ints = {11, 12, 13})
-	@DisplayName("핀 수가 최소 수 보다 작을 경우")
-	void maximumException(int pin) {
-		assertThatExceptionOfType(IllegalArgumentException.class)
-			.isThrownBy(() -> new Pin(pin));
-	}
-
-}
